Convert driver supported capabilities to compute node provider traits
This builds on the ProviderTree work in the compute driver and
resource tracker to take the supported capabilities from a driver and
turn those into traits on the compute node resource provider.  If a
standard trait exists corresponding to the capability then that is
used, otherwise the trait is formed by upper-casing the capability
name and prefixing it with "CUSTOM_COMPUTE_", e.g.

    CUSTOM_COMPUTE_HAS_IMAGECACHE

This is a simple way to expose in a REST API (Placement in this case)
what a compute node, via its driver, supports.

This is also something easy that we can do in lieu of a full-blown
compute capabilities REST API in nova, which we've talked about for
years but never actually done anything about.

We can later build on this to add a request filter which will mark
certain types of boot-from-volume requests as requiring specific
capabilites, like for volume multiattach and tagged devices.

Any traits provided by the driver will be automatically added during
startup or a periodic update of a compute node:

    https://pasteboard.co/I1Tbzp2.png

Similarly any traits later retracted by the driver will be
automatically removed.

However any custom traits which are removed from the Placement API via
the admin will not be reinstated until the compute service's provider
cache is reset.

The new call graph is shown in this sequence diagram:

    https://pasteboard.co/I25qICd.png

Co-Authored-By: Adam Spiers <aspiers@suse.com>
Depends-On: https://review.openstack.org/576595/

Related to blueprint placement-req-filter
Related to blueprint expose-host-capabilities

Change-Id: I15364d37fb7426f4eec00ca4eaf99bec50e964b6