WIP: Detect that SEV is required and enable iommu for devices
This is still WIP - the _sev_required check needs to be changed to
return true iff:

  a) the supports_amd_sev instance variable in the driver is
     true, *and*

  b) the allocations fed to spawn() show that one slot of
     the MEM_ENCRYPTION_CONTEXT resource inventory has been
     allocated to the guest.

When MEM_ENCRYPTION_CONTEXT is included in the inventory allocations
for the instance, and the host is capable of SEV, we know that SEV
functionality is required, so the guest config needs to be tweaked
accordingly in various ways.  This change implements the first of
those tweaks, which is to set all virtio devices' "iommu" attributes
to "on".  Subsequent commits will add the other tweaks.

blueprint: amd-sev-libvirt-support
Change-Id: Ie54fca066f3333d1d5d18a2c0e8f6c7d5042490b