Provide HW_CPU_X86_AMD_SEV trait when SEV is supported
Add the final piece of the SEV detection code to init_host() in the
libvirt driver, by checking that:

   a) 'sev' parameter of the kvm-amd kernel module is enabled, and

   b) the <features> element returned from libvirt's
      virConnectGetDomainCapabilities API contains an

         <sev supported='yes'>

      element.  This second check is achieved by utilising the domain
      capability retrieval and parsing code which was added in
      I4c35b6a27db05349213429422ffe62adb09bd921.

When both checks pass, provide the standard trait HW_CPU_X86_AMD_SEV
(which is newly available in os-traits 0.13.0) on the resource
provider for the compute host.  This indicates the ability of the
compute host to boot instances with AMD SEV (Secure Encrypted
Virtualization) enabled.

Since the functional test setup is near identical to that already used
in test_shared_resource_provider.py, extract it into a new base class
called LibvirtProviderUsageBaseTestCase in a new file:

    nova/tests/functional/libvirt/integrated_helpers.py

and reuse that for the new tests.

blueprint: amd-sev-libvirt-support
Change-Id: I2b41f1cce0af8b9d36b74a4663aa4ff227e17cc6