Track inventory for new MEM_ENCRYPTION_CONTEXT resource class
Track compute node inventory for the new MEM_ENCRYPTION_CONTEXT
resource class (added in os-resource-classes 0.4.0) which represents
the number of guests a compute node can host concurrently with memory
encrypted at the hardware level.

Initially this is required because on AMD SEV-capable hardware, the
memory controller has a fixed number of slots for holding encryption
keys, one per guest.  Typical hardware only has 15 slots, thereby
limiting the number of SEV guests which can be run concurrently to 15.
nova needs to track how many slots are available and used in order to
avoid attempting to exceeding that limit in the hardware.

Work is in progress to allow QEMU and libvirt to expose the number of
slots available on SEV hardware; however until this is finished and
released, it will not be possible for nova to programatically detect
the correct value with which to populate the MEM_ENCRYPTION_CONTEXT
inventory.  So add a new CONF.libvirt.num_memory_encrypted_guests
configuration option as a stop-gap, allowing the cloud operator to
provide this value manually, and then use it to populate the
inventory.

blueprint: amd-sev-libvirt-support
Change-Id: Ib1304e65f1a203e6aed66e530d6e573c36165cd7